Problem: test_monitor depends on testutil_security
Solution: move general monitoring test utilities to new testutil_monitoring
