+++ tests/CMakeLists.txt
@@ -186,7 +186,7 @@ foreach(test ${tests})
     add_executable(${test} ${test}.cpp
       "testutil_security.hpp")
   else()
-    add_executable(${test} ${test}.cpp "testutil.hpp" "testutil_unity.hpp")
+    add_executable(${test} ${test}.cpp "testutil.hpp" "testutil_unity.hpp" "testutil_monitoring.hpp")
   endif()
   if(WIN32)
     # This is the output for Debug dynamic builds on Visual Studio 6.0
