+++ tests/test_monitor.cpp
@@ -28,7 +28,7 @@
 */
 
 #include "testutil.hpp"
-#include "testutil_security.hpp"
+#include "testutil_monitoring.hpp"
 
 #include "testutil_unity.hpp"
